Fix an issue where types that are binary coercible to other types wouldn't use those type comparators for the skip list, leading to filters on columns of those types forcing to scan through the whole table.
This is a problem with VARCHAR columns, for example, where they get compared using `varchar_ops` from the TEXT type since they're binary-compatible, but CStore doesn't detect that.

See definition of `IsBinaryCoercible` for a full list of types this impacts. Full fix will require regenerating existing CStore files to include the min-max values for VARCHARs in the block's header.